refactor(acp-bridge): lift bridge types (#4175 PR 22b checkpoint 2)
Pure type lift of the bridge's public TypeScript surface — no behavior
or runtime change.

What moves
- 11 type declarations (~520 LOC) from `httpAcpBridge.ts:101-620` →
  `packages/acp-bridge/src/bridgeTypes.ts`:
  - `BridgeSpawnRequest`, `BridgeSession`, `BridgeRestoreSessionRequest`
  - `BridgeSessionState` (alias of ACP `LoadSessionResponse | ResumeSessionResponse`)
  - `BridgeRestoredSession`, `BridgeSessionSummary`, `SessionMetadataUpdate`
  - `BridgeClientRequestContext`, `BridgeHeartbeatResult`, `BridgeHeartbeatState`
  - `HttpAcpBridge` interface (the bridge's public facade — ~30 methods
    + 2 getter properties)

`bridgeTypes.ts` imports
- `ApprovalMode` from `@qwen-code/qwen-code-core` (status types reference it)
- ACP wire types (`CancelNotification`, `LoadSessionResponse`,
  `PromptRequest`, `PromptResponse`, `RequestPermissionResponse`,
  `ResumeSessionResponse`, `SetSessionModelRequest`,
  `SetSessionModelResponse`) from `@agentclientprotocol/sdk`
- `BridgeEvent` + `SubscribeOptions` from local `./eventBus.js`
- All `Serve*Status` types from local `./status.js` (also part of
  acp-bridge after checkpoint 1)

Wrappers
- `httpAcpBridge.ts` lines 101-620 replaced with a single `import type` +
  `export type` re-export block; the local `BridgeClient` class +
  `createHttpAcpBridge` factory continue to reference these names
  through the import binding
- Cleaned up imports from `@agentclientprotocol/sdk` and `./status.js` —
  `LoadSessionResponse`, `PromptResponse`, `ResumeSessionResponse`,
  `SubscribeOptions`, and 5 `ServeWorkspace*Status` types are no longer
  used inline in `httpAcpBridge.ts` (all consumed via lifted types)
- Added `./bridgeTypes` subpath export to `acp-bridge/package.json`

Backward compatibility
- All existing imports of `from './httpAcpBridge.js'` keep resolving
  (server.ts:31, runQwenServe.ts:16, workspaceAgents.ts:21,
  workspaceMemory.ts:19, index.ts:90-97)
- Zero `/capabilities`, route, SDK, or behavior changes

Verification
- `cd packages/core && npx tsc --build` clean
- `cd packages/acp-bridge && npx tsc --build` clean
- `cd packages/cli && npx tsc --noEmit` clean

Remaining PR 22b checkpoints
- 22b/3: BridgeOptions + DaemonStatusProvider injection seam
- 22b/4: BridgeClient class
- 22b/5: defaultSpawnChannelFactory
- 22b/6: createHttpAcpBridge factory (~2240 LOC of moves) — the bulk
- 22b/7: serve/daemonStatusProvider.ts + runQwenServe wire
- 22b/8: move bridge tests (5064 lines)
- 22b/9: shrink httpAcpBridge.ts to final re-export shim
